pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-system"
version = "1.0.0"
description = "BrainCore + DLC framework with production hardening"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Brain AI Systems" }]
license = { text = "Proprietary" }

dependencies = [
  "packaging>=23.2",
]

[project.optional-dependencies]
crypto = [
  "cryptography>=42.0.0",
]
observability = [
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.22.0",
  "opentelemetry-sdk>=1.22.0",
]
server = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
]
logging = [
  "python-json-logger>=2.0.7",
]
config = [
  "watchdog>=4.0.0",
  "python-consul2>=0.1.5",
]
ha = [
  "redis>=5.0.0",
]
docs = [
  "sphinx>=7.2.0",
]

[project.scripts]
brain = "brain_system.cli:main"
brain-ui = "brain_system.ui:main"

[tool.setuptools]
packages = ["brain_system"]
py-modules = ["Bain", "BainDLC"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["brain_system"]

[tool.coverage.report]
fail_under = 80
show_missing = true
